src/requirements.txt
Flask
requests
PySocks
waitress
psycopg2-binary
# datetime
